requirements.txt
fastapi
uvicorn
psycopg2-binary
geopandas
owslib
python-dotenv
shapely
sqlalchemy